MySQL tables export
These are tables exported from MySQL. They are filled with the database information provided by the professor, along with users that have varying privileges (used for testing). Some data modification had to be done, such as changing the date format of each row/entry to fit the DATE attribute syntax, but this is the closest version to the final product.